+++ examples/server.php
@@ -8,17 +8,19 @@ use Doctrine\DBAL\DriverManager;
 use Flow\Driver\AmpDriver;
 use Flow\Driver\FiberDriver;
 use Flow\Driver\ReactDriver;
+use Flow\Driver\RevoltDriver;
 use Flow\Driver\SwooleDriver;
 use Flow\Examples\Transport\DoctrineIpTransport;
 use Flow\Flow\Flow;
 use Flow\Flow\TransportFlow;
 use Flow\IpStrategy\MaxIpStrategy;
 
-$driver = match (random_int(1, 4)) {
+$driver = match (random_int(1, 5)) {
     1 => new AmpDriver(),
     2 => new ReactDriver(),
     3 => new FiberDriver(),
-    4 => new SwooleDriver(),
+    4 => new RevoltDriver(),
+    5 => new SwooleDriver(),
 };
 printf("Use %s\n", $driver::class);